Moved pub recycler list to child fragment PubListFragment
