+++ app/src/main/res/layout/fragment_pub_view.xml
@@ -59,14 +59,16 @@
 
     </android.support.constraint.ConstraintLayout>
 
-    <android.support.v7.widget.RecyclerView
-        android:id="@+id/recycler_pub_view"
+    <FrameLayout
+        android:id="@+id/frame_pubs"
         android:layout_width="0dp"
-        android:layout_height="wrap_content"
+        android:layout_height="0dp"
+        app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/layout_search" />
+        app:layout_constraintTop_toBottomOf="@+id/layout_search">
+
+    </FrameLayout>
 
 
 </android.support.constraint.ConstraintLayout>
